+++ src/main/java/com/theone/nailtherapyspring/service/Service.java
@@ -1,5 +1,6 @@
-package com.theone.nailtherapyspring.models;
+package com.theone.nailtherapyspring.service;
 
+import com.theone.nailtherapyspring.models.AbstractEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotNull;
